Added java project. Compile & run this java code
Directory java/ contains the code that will be executed by the JVM at
launch. This is meant to be a "server-side" code (not user-supplied
code) that will govern what and how is executed under pgj. It is the pgj
code, Java side.

Changed build scripts and C code, as well as README.md, to reflect these
changes. Build process now expects that java code is built first, as the
.jar artifact will be created as part of the C extension.